# chore(libzkp): upgrade to v0.10.5, speed up by avoiding some unnecessary code hashing