Since code_size is stored inside the Account struct, we can use it instead of reading full bytecode.

CCC avg time(rust only) of recent blocks: 

before this: deserialize_elapsed: 144.32ms, total_elapsed: 599.72ms
after this: deserialize_elapsed: 143.04ms, total_elapsed: 357.58ms

original commit: https://github.com/scroll-tech/zkevm-circuits/pull/1256/commits/30ed394f24c73a8d1537805f30a43877f1096f4d